at 71% class coverage and 63% line coverage see #18
