+++ src/test/java/io/nats/bridge/example/a/BridgeManagerProtoMain.java
@@ -11,46 +11,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package io.nats.bridge.example.service.b;
+package io.nats.bridge.example.a;
 
+import io.nats.bridge.MessageBridge;
 import io.nats.bridge.MessageBus;
-import io.nats.bridge.example.service.a.ServiceAUtil;
 
-import java.util.Arrays;
-import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 
 //TODO turn this into a test.
 // See https://github.com/nats-io/nats-jms-mq-bridge/issues/16
-public class JMSHelloWorldClient {
+public class BridgeManagerProtoMain {
 
     public static void main(String... args) {
         try {
+            final MessageBus messageBusSource = ServiceAUtil.getMessageBusNats();
+            final MessageBus messageBusDestination = ServiceAUtil.getMessageBusJms();
+            final MessageBridge messageBridge = new MessageBridge(messageBusSource, messageBusDestination);
+
             final AtomicBoolean stop = new AtomicBoolean(false);
-            final MessageBus messageBus = ServiceBUtil.getMessageBusJms();
-            final List<String> names = Arrays.asList("Rick", "Tom", "Chris", "Paul", "Noah", "Lucas");
+
             Runtime.getRuntime().addShutdownHook(new Thread(() -> stop.set(true)));
 
-            int count = 0;
             while (true) {
-                Thread.sleep(1);
-                if (stop.get()) {
-                    messageBus.close();
-                    break;
-                }
-                final int index = count;
-                names.forEach(name -> {
-                    System.out.println("Sending: " + name + index);
-                    messageBus.request(name + index, s -> System.out.println("Received: " + s));
-                });
-                count++;
-                Thread.sleep(100);
-                messageBus.process();
+                if (stop.get()) break;
+                Thread.sleep(10);
+                messageBridge.process();
             }
+            messageBridge.close();
 
         } catch (Exception ex) {
             ex.printStackTrace();
         }
     }
+
 }
