+++ src/test/java/io/nats/bridge/example/a/ServiceAUtil.java
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package io.nats.bridge.example.service.a;
+package io.nats.bridge.example.a;
 
 import io.nats.bridge.MessageBus;
 import io.nats.bridge.jms.support.JMSMessageBusBuilder;
@@ -22,16 +22,16 @@ import io.nats.client.Options;
 import java.io.IOException;
 import java.util.UUID;
 
-public class ServiceATestUtil {
+public class ServiceAUtil {
     static MessageBus getMessageBusJms() {
-        final String queueName = "dynamicQueues/A1QueueTest-1";
+        final String queueName = "dynamicQueues/A1Queue";
         final JMSMessageBusBuilder jmsMessageBusBuilder = new JMSMessageBusBuilder().withDestinationName(queueName);
         final MessageBus messageBus = jmsMessageBusBuilder.build();
         return messageBus;
     }
 
     static MessageBus getMessageBusNats() throws IOException, InterruptedException {
-        final String subject = "a1-subject-test-1";
+        final String subject = "a1-subject";
 
         final Options options = new Options.Builder().
                 server("nats://localhost:4222").
