+++ src/test/java/io/nats/bridge/example/b/ServiceBNatsServer.java
@@ -11,41 +11,43 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package io.nats.bridge.example.service.a;
+package io.nats.bridge.example.b;
 
+import io.nats.bridge.Message;
 import io.nats.bridge.MessageBus;
+import io.nats.bridge.StringMessage;
 
-import java.util.Arrays;
-import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.Consumer;
 
 
 //TODO turn this into a test.
-// See https://github.com/nats-io/nats-jms-mq-bridge/issues/16
-public class NatsHelloWorldClient {
+public class ServiceBNatsServer {
 
     public static void main(String... args) {
         try {
+
             final AtomicBoolean stop = new AtomicBoolean(false);
-            final MessageBus messageBus = ServiceAUtil.getMessageBusNats();
-            final List<String> names = Arrays.asList("Rick", "Tom", "Chris", "Paul", "Noah", "Lucas");
+            final MessageBus messageBus = ServiceBUtil.getMessageBusNats();
+
             Runtime.getRuntime().addShutdownHook(new Thread(() -> stop.set(true)));
 
-            int count = 0;
             while (true) {
-                Thread.sleep(1);
                 if (stop.get()) {
                     messageBus.close();
                     break;
                 }
-                final int index = count;
-                names.forEach(name -> {
-                    System.out.println("Sending: " + name + index);
-                    messageBus.request(name + index, s -> System.out.println("Received: " + s));
+                final Optional<Message> receive = messageBus.receive();
+                receive.ifPresent(message -> {
+
+                    StringMessage stringMessage = (StringMessage) message;
+                    System.out.println("Handle message " + stringMessage.getBody());
+                    message.reply(new StringMessage("Hello " + stringMessage.getBody()));
                 });
-                count++;
-                Thread.sleep(1000);
+
+
+                Thread.sleep(10);
+                messageBus.process();
             }
 
         } catch (Exception ex) {
